Revert "Add CI pipeline (unit tests + linter)"
