--- .github/workflows/node.js.yml
@@ -1,26 +1,0 @@
-name: Push (lint + unit tests)
-
-on: [push]
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-
-    strategy:
-      matrix:
-        node-version: [16.x]
-
-    steps:
-      - uses: actions/checkout@v3
-      - name: Use Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v3
-        with:
-          node-version: ${{ matrix.node-version }}
-      - name: npm install, build, and test
-        run: |
-          npm install
-          npm run build
-          npm run lint
-          npm run test
-        env:
-          CI: true
