+++ packages/app/package.json
@@ -15,7 +15,7 @@
   "homepage": "https://github.com/martinascharrer/vuensight#readme",
   "scripts": {
     "serve": "vue-cli-service serve",
-    "build": "rm -rf dist/ && echo 'Building app...' && vue-cli-service build --dest dist/app && echo 'Building server...' && cd server && tsc -p tsconfig.pkg.json",
+    "build": "rm -rf dist/ && vue-cli-service build --dest dist/app && cd server && tsc -p tsconfig.pkg.json",
     "build:watch": "vue-cli-service build --watch",
     "test": "vue-cli-service test:unit",
     "lint": "vue-cli-service lint",
@@ -47,7 +47,6 @@
     "@vue/eslint-config-airbnb": "^6.0.0",
     "@vue/eslint-config-typescript": "^10.0.0",
     "@vue/test-utils": "^2.0.0-rc.17",
-    "@vue/vue3-jest": "^27.0.0",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-vue": "^8.5.0",
     "sass": "^1.49.9",
