+++ packages/app/src/components/base/BaseCheckIcon.vue
@@ -21,7 +21,7 @@
 <script lang="ts">
 import { defineComponent, PropType } from 'vue';
 
-import { Color } from '@/types/color';
+import { Color } from '@/types';
 
 export default defineComponent({
   props: {
