+++ packages/app/src/components/base/BaseDelimiter.vue
@@ -12,7 +12,7 @@
 <script lang="ts">
 import { defineComponent, PropType } from 'vue';
 
-import { Color } from '@/types/color';
+import { Color } from '@/types';
 
 export default defineComponent({
   name: 'BaseDelimiter',
