+++ packages/app/src/components/base/BaseList.vue
@@ -18,7 +18,7 @@
 <script lang="ts">
 import { defineComponent, PropType } from 'vue';
 
-import { Color } from '@/types/color';
+import { Color } from '@/types';
 
 export default defineComponent({
   name: 'BaseDelimiter',
