+++ packages/app/src/services/parser.ts
@@ -1,7 +1,4 @@
-import { VueComponent } from '@vuensight/types';
-
-export const get = ():Promise<VueComponent[]> => fetch('http://localhost:4444/parse-result')
-  .then((result) => result.json());
+export const get = ():Promise<any> => fetch('http://localhost:4444/parse-result').then((result) => result.json());
 
 export default {
   get,
