--- packages/app/src/types/color.ts
@@ -1,1 +1,0 @@
-export type Color = 'mint' | 'red' | 'purple' | 'light-mint' | 'light-red' | 'light-purple';
