--- packages/app/src/types/force.ts
@@ -1,11 +1,0 @@
-import { VueComponent } from '@vuensight/types';
-
-export type Link = {
-    source: string,
-    target: string,
-}
-
-export type ForceLayout = {
-    nodes: VueComponent[],
-    links: Link[],
-}
