+++ packages/app/tsconfig.json
@@ -36,7 +36,6 @@
     "tests/**/*.tsx"
   ],
   "exclude": [
-    "node_modules",
-    "server/**/*"
+    "node_modules"
   ]
 }
