+++ packages/parser/package.json
@@ -27,11 +27,7 @@
   "homepage": "https://github.com/martinascharrer/vuensight#readme",
   "jest": {
     "preset": "ts-jest",
-    "testEnvironment": "node",
-    "moduleNameMapper": {
-      "dependency-cruiser/config-utl/extract-ts-config": "dependency-cruiser/src/config-utl/extract-ts-config",
-      "dependency-cruiser/config-utl/extract-webpack-resolve-config": "dependency-cruiser/src/config-utl/extract-webpack-resolve-config"
-    }
+    "testEnvironment": "node"
   },
   "dependencies": {
     "dependency-cruiser": "^11.4.1",
